+++ examples/dual-region_vault-with-dr/main.tf
@@ -17,6 +17,12 @@ locals {
     }
   }
 
+  # General locals
+  common_tags = {
+    Creator               = "CyberArk PAMonCloud via Terraform"
+    Tf_Plan_Creation_Date = plantimestamp()
+  }
+
   # Primary Vault locals
   vault_instance_name            = "[PAMonCloud_TF] Primary Vault"
   vault_instance_type            = "m5.2xlarge"
@@ -33,11 +39,46 @@ locals {
 provider "aws" {
   region = local.regions.main_region.name
   alias  = "main"
+
+  default_tags {
+    tags = merge(
+      local.common_tags,
+      {
+        Region_Role = "Primary"
+      }
+    )
+  }
 }
 
 provider "aws" {
   region = local.regions.dr_region.name
   alias  = "dr"
+
+  default_tags {
+    tags = merge(
+      local.common_tags,
+      {
+        Region_Role = "DR"
+      }
+    )
+  }
+}
+
+################################################################################
+# deploy_prerequisites Module
+################################################################################
+module "deploy_prep_main" {
+  source = "../../modules/deploy_prerequisites"
+  providers = {
+    aws = aws.main
+  }
+}
+
+module "deploy_prep_dr" {
+  source = "../../modules/deploy_prerequisites"
+  providers = {
+    aws = aws.dr
+  }
 }
 
 ################################################################################
@@ -63,23 +104,6 @@ module "pam_network_dr" {
   network_type               = local.regions.dr_region.network_type
   users_access_cidr          = local.regions.dr_region.users_access_cidr
   administrative_access_cidr = local.regions.dr_region.administrative_access_cidr
-}
-
-################################################################################
-# deploy_prerequisites Module
-################################################################################
-module "deploy_prep_main" {
-  source = "../../modules/deploy_prerequisites"
-  providers = {
-    aws = aws.main
-  }
-}
-
-module "deploy_prep_dr" {
-  source = "../../modules/deploy_prerequisites"
-  providers = {
-    aws = aws.dr
-  }
 }
 
 ################################################################################
@@ -122,6 +146,7 @@ module "vault_instance" {
   providers = {
     aws = aws.main
   }
+  deployment_identifier          = module.deploy_prep_main.deployment_uid
   instance_name                  = local.vault_instance_name
   instance_type                  = local.vault_instance_type
   key_name                       = var.key_name
@@ -151,6 +176,7 @@ module "vault_dr_instance" {
   providers = {
     aws = aws.dr
   }
+  deployment_identifier          = module.deploy_prep_dr.deployment_uid
   instance_name                  = local.vaultdr_instance_name
   instance_type                  = local.vaultdr_instance_type
   key_name                       = var.key_name