+++ modules/component/iam.tf
@@ -79,7 +79,7 @@ resource "aws_iam_role_policy" "instance_kms_policy" {
 #### IAM Roles
 
 resource "aws_iam_role" "instance_role" {
-  name               = "PAMonCloud_TF_${var.component}_Role_${random_string.deployment_uid.result}"
+  name               = "PAMonCloud_TF_${var.component}_Role_${local.deployment_uid}"
   path               = "/"
   assume_role_policy = data.aws_iam_policy_document.instance_assume_role_policy.json
 }
@@ -87,6 +87,6 @@ resource "aws_iam_role" "instance_role" {
 #### IAM Instance Profiles
 
 resource "aws_iam_instance_profile" "instance_profile" {
-  name = "PAMonCloud_TF_${var.component}_InstanceProfile_${random_string.deployment_uid.result}"
+  name = "PAMonCloud_TF_${var.component}_InstanceProfile_${local.deployment_uid}"
   role = aws_iam_role.instance_role.name
 }