+++ modules/pam_network/variables.tf
@@ -28,3 +28,36 @@ variable "network_type" {
     error_message = "Invalid network type. Valid options are: 'privatelink', 'nat'"
   }
 }
+
+variable "vpn_customer_gateway_address" {
+  description = "Public IP address of the remote network"
+  type        = string
+  default     = ""
+}
+
+variable "vpn_external_vault_cidr" {
+  description = "IPv4 address range of the remote network hosting the external Vault"
+  type        = string
+  default     = ""
+  validation {
+    condition     = var.vpn_external_vault_cidr != var.pam_vpc_cidr
+    error_message = "Overlapping CIDRs. vpn_external_vault_cidr can't overlap with pam_vpc_cidr"
+  }
+}
+
+variable "log_group_arn" {
+  description = "ARN of the CloudWatch log group."
+  type        = string
+  default     = ""
+  validation {
+    condition = (
+      var.log_group_arn == "" ||
+      can(regex(
+        "^arn:(aws|aws-us-gov):logs:[a-z0-9-]+:[0-9]{12}:log-group:[A-Za-z0-9_./-]+(?::\\*)?$",
+        var.log_group_arn
+        )
+      )
+    )
+    error_message = "The value must be a valid CloudWatch log group ARN (e.g. arn:aws:logs:region:123456789012:log-group:my-log-group or with :* suffix)."
+  }
+}