+++ modules/vault_dr/main.tf
@@ -14,6 +14,7 @@ locals {
   success_signal       = local.component_data[local.component].success_signal
   log_streams          = local.component_data[local.component].log_streams
   user_data            = local.component_data[local.component].user_data
+  deployment_uid       = var.deployment_identifier == "" ? random_string.deployment_uid.result : var.deployment_identifier
 }
 
 resource "null_resource" "always_recreate" {